chore: move mergeUserWebpackConfig to webpack.common
