+++ packages/core/config/webpack.common.js
@@ -8,6 +8,8 @@ const MiniCssExtractPlugin = require('mini-css-extract-plugin')
 const CopyPlugin = require('copy-webpack-plugin')
 const Dotenv = require('dotenv-webpack')
 const { getConfig, getPaths } = require('@redwoodjs/internal')
+const { existsSync } = require('fs')
+const merge = require('webpack-merge')
 
 const redwoodConfig = getConfig()
 const redwoodPaths = getPaths()
@@ -196,4 +198,19 @@ module.exports = (webpackEnv) => {
         : (info) => path.resolve(info.absoluteResourcePath).replace(/\\/g, '/'),
     },
   }
+}
+
+module.exports['mergeUserWebpackConfig'] = (mode, baseConfig) => {
+  const redwoodPaths = getPaths()
+  const hasCustomConfig = existsSync(redwoodPaths.web.webpack)
+  if (!hasCustomConfig) {
+    return baseConfig
+  }
+  const userWebpackConfig = require(redwoodPaths.web.webpack)
+
+  if (typeof userWebpackConfig === 'function') {
+    return userWebpackConfig(baseConfig, {mode})
+  }
+
+  return merge(baseConfig, userWebpackConfig)
 }
