+++ packages/core/config/webpack.production.js
@@ -1,6 +1,5 @@
 const webpackConfig = require('./webpack.common')
-const { mergeUserWebpackConfig } = require('./utils')
-
+const { mergeUserWebpackConfig } = webpackConfig
 const baseConfig = webpackConfig('production')
 
 module.exports = mergeUserWebpackConfig('production', baseConfig)
